接続確認待ち時間を実装.

+++ 11/DX12Tutorial11/Src/Main.cpp
@@ -264,7 +264,7 @@ bool Render()
 */
 void Update(double delta)
 {
-	UpdateGamePad(delta);
+	UpdateGamePad(static_cast<float>(delta));
 	GamePad& gamepad = GetGamePad(GamePadId_1P);
 
 	Graphics::Graphics& graphics = Graphics::Graphics::Get();
